Move fileStore file in a proper dir
