+++ src/__test__/fileStore.test.js
@@ -1,11 +1,11 @@
 import request from 'supertest';
 import express from 'express';
-import fileStore from '../fileStore';
+import fileStore from '../fileStore/fileStore';
 import {
   MemoryFileBackend,
   DiskFileBackend,
   S3FileBackend,
-} from '../fileStoreBackend';
+} from '../fileStore/fileStoreBackend';
 import path from 'path';
 import fs from 'fs';
 import tempy from 'tempy';
