+++ src/fileStore/fileStoreBackend.js
@@ -6,7 +6,7 @@ import path from 'path';
 import fs from 'fs';
 import { Duplex } from 'stream';
 
-import { uid } from './uid.js';
+import { uid } from '../uid.js';
 
 const bufferToStream = (buffer) => {
   let stream = new Duplex();
